Fixed connfd type error
The connfd should be an int, not unsigned long because this prevents us
from capturing net_accept errors

+++ lib/protocol_server.c
@@ -246,7 +246,7 @@ void *
 proto_server_rpc_listen(void *arg)
 {
   int fd = Proto_Server.RPCListenFD;
-  unsigned long connfd;
+  int connfd;
   pthread_t tid;
   
   if (net_listen(fd) < 0) {
